Temporary workaround for 127.0.0.1:4381
Add checks for `err.code` value of `ECONNREFUSED`. It notifies user if their Spotify is either closed or unreachable on port 4381, and advises them to close it + wait 10 seconds before relaunching to ensure it'll fall back to :4381. The process silently checks for port 4381 every 5 seconds instead of 1.5 seconds, and restores the flow back to 1.5 seconds once it detects the port. Partially references issue #24 with own touch added to it.